packet: fix pkt data offset may larger than 256 on xdp
When running xdp nic on kernel 5.10, it is observed that offset extract from
xdp_desc was larger then 700, resulting data offset of mbuf being larger then
256. Therefore, releated offsets was changed from uint8_t to uint16_t to
avoid potential overflow.

+++ include/packet.h
@@ -38,9 +38,9 @@ struct packet {
 
 	uint16_t flags;
 	uint8_t  wid;
-	uint8_t l2_off;
-	uint8_t l3_off;
-	uint8_t l4_off;
+	uint16_t l2_off;
+	uint16_t l3_off;
+	uint16_t l4_off;
 
 	/*
 	 * The tcp payload offset; unlike other offsets, it's not fixed.
